Fix MoEFFN autocast dtype mismatch; record AMX benchmark on 8480 ES
Bug (found via CPU bf16 benchmarking): MoEFFN dispatch did
`out.index_add_(0, tok, expert_out)` where `out` inherits the input-activation
dtype but `expert_out` comes from a Linear that autocast runs in bf16. When the
MoE input arrives fp32 under autocast (the CPU path), the scatter crashed on a
scalar-type mismatch. The CUDA training path dodged it because dtypes happened
to align. Fix: cast the scatter source to the accumulator dtype (no-op when
they already match -> zero change to the working CUDA/native paths). +2
regression tests (MoE and dense forward under CPU bf16 autocast).

Also recorded in roadmap: benchmarked on the on-hand Sapphire Rapids 8480 ES.
AMX is NOT engaged under stock Windows torch+cpu (bf16 GEMM ~3 TFLOPS =
AVX-512-BF16, not AMX); distill_tiny ran 85 tok/s. Realising AMX needs IPEX +
likely Linux + retail clocks + 8 memory channels -- it's a config project, not
plug-and-play. Validate AMX is actually firing before assuming the speedup.

+++ mythouro/main.py
@@ -755,11 +755,15 @@ class MoEFFN(nn.Module):
             tok = token_index[sel]                                # (M,)
             gate = flat_gate[sel]                                 # (M, 1)
             expert_out = self.routed_experts[eid](flat[tok]) * gate
-            out.index_add_(0, tok, expert_out)
+            # `.to(out.dtype)` keeps the scatter dtype-consistent under
+            # autocast (where the expert Linear emits bf16 but `out` may be the
+            # fp32 input activation). No-op when dtypes already match — i.e. on
+            # the native-bf16 / fp32 paths.
+            out.index_add_(0, tok, expert_out.to(out.dtype))
 
         # shared experts always fire for every token
         for shared in self.shared_experts:
-            out = out + shared(flat)
+            out = out + shared(flat).to(out.dtype)
 
         return out.view(B, T, D)
